clim-examples: add new regression test demo (patterns and designs)
We add rectangular-tiles in various settings (we assume it is a pattern) and a
stencil buffer (disabled because it breaks). Also more indexed designs and such.

We also create an artificial uniform-ink-pattern object for testing. The purpose
of that is to be able to see shape and placement of pattern drawing without
worrying that ink transformatino is bogus (we use +blue+ which covers
+everywhere+). Usage:

Keypress 1: use draw-pattern
Keypress 2: use draw-pattern with uniform ink
Keypress 3: use draw-design
Keypress 4: use draw-design with uniform ink
Keypress 5: use draw-rectangle with pattern bound to ink

Keypress space: dummy command (triggers redisplay)
Keypress r: start new demo and close this frame (restart)
Keypress q: exit demo

1-5 commands map over all sheets and first call redisplay on them and then
repaint. That assures, that all is redrawn. If in doubt press r (currently used
draw function is preserved across restarts).